% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayover.R
\name{arrayover}
\alias{arrayover}
\title{Retorna tokens comuns entre dois arrays informados nos argumentos}
\usage{
arrayover(vetor1, vetor2)
}
\value{
Retorna um vetor de strings contendo os tokens em comum entre os dois conjuntos recebidos
como argumentos
}
\description{
Esta função recebe dois arrays de tokens e retorna um novo array contendo
todos os tokens comuns. Isto é, retorna a intercessão dos conjuntos passados como argumentos.
}
\author{
Jones Quito
}
