% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Calcula a sobreposição de dois vetores}
\usage{
overlap(vetor1, vetor2)
}
\arguments{
\item{vetor1}{vetor de tokens}

\item{vetor2}{vetor de tokens}
}
\value{
Retorna um inteiro indicando a sobreposição dos elementos de vetor1 e vetor2
}
\description{
Função que calcula a sobreposição de dois vetores passados como argumento
}
\author{
Jones Quito
}
