% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Calcula a similaridade de duas strings}
\usage{
similarity(s1, s2, q, simFunc)
}
\arguments{
\item{s1}{é a primeira de duas strings a ser comparadas}

\item{s2}{é a segunda de duas strings a ser comparadas}

\item{q}{o tamanho dos tokens a ser gerados para calculo de similaridade - dois(2) é um tamanho bom}

\item{simFunc}{um inteiro que representa a função de similaridad a ser usada. As opções são:
1 - Similaridade de Jaccard
2 - Similaridade Dice
3 - Similaridade Cosine}
}
\value{
Retorna um valor float entre 0 e 1 representando a porcentagem de similaridade entre as strings.
}
\description{
Esta função retorna um valor de ponto flutuante que repesenta a similaridade entre
duas strings informadas como argumentos. A similaridade é calculada de acordo com o algorítimo
informado com argumento.
}
\examples{
texto1 <- 'Felipe de Sousa'
texto2 <- 'Felipe de Souza'

sim <- similarity(texto1, texto2, 3, 1)

}
\author{
Jones Quito
}
